rag_chatbot/requirements.txt
streamlit>=1.32.0
google-genai>=1.0.0
pinecone>=5.0.0
pdfplumber>=0.10.0
langchain-text-splitters>=0.2.0
python-dotenv>=1.0.0
